feat: new storage implementation
Closes #8
